move start to own command to allow running it without arguments
